Agrego logo y refactor de directorios
