+++ src/app/components/marketplace/marketplace.component.scss
@@ -1,6 +1,6 @@
 /* ===== MARKETPLACE REDESIGN - PROFESSIONAL E-COMMERCE ===== */
 
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .marketplace-container {
   min-height: 100vh;
