+++ src/app/components/my-products/my-products.component.scss
@@ -1,6 +1,6 @@
 /* ===== MY PRODUCTS REDESIGN - PROFESSIONAL DASHBOARD ===== */
 
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .my-products-container {
   min-height: 100vh;
