+++ src/app/components/my-products/my-products.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
 import { Subject, takeUntil } from 'rxjs';
-import { ProductService } from '../core/services/product.service';
-import { AuthService } from '../core/services/auth.service';
-import { Product } from '../core/models/product.model';
+import { ProductService } from '../../core/services/product.service';
+import { AuthService } from '../../core/services/auth.service';
+import { Product } from '../../core/models/product.model';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
