+++ src/app/components/product-detail/product-detail.component.scss
@@ -1,6 +1,6 @@
 /* ===== PRODUCT DETAIL REDESIGN - PROFESSIONAL E-COMMERCE ===== */
 
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .product-detail-container {
   min-height: 100vh;
