+++ src/app/components/quotation-dialog/quotation-dialog.component.scss
