+++ src/app/components/quotation-dialog/quotation-dialog.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { Product } from '../core/models/product.model';
+import { Product } from '../../core/models/product.model';
 
 export interface QuotationData {
   product: Product;
