+++ src/app/core/constants/images.constants.ts
@@ -1,0 +1,3 @@
+export const logo = {
+    logoAgro: 'assets/images/logo_agro_mas.png',
+};
